Handle UserNotFoundException for ProfileAvatarController endpoints
